pyproject.toml
[tool.poetry]
name = "rag-chatbot"
version = "0.1.0"
description = "A chatbot utilising Retrieval Augmented Generation."
authors = ["Renat Oosthuizen"]
readme = "README.md"
packages = [{include = "rag-chatbot"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
llama-index-core = "^0.10.5"
python-dotenv = "^1.0.1"
pandas = "^2.2.0"
pyarrow = "^15.0.0"
llama-index = "^0.10.5"
pip = "^24.0"
llama-index-readers-file = "^0.1.3"
llama-index-llms-openai = "^0.1.2"
docx2txt = "^0.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
